Allow for integration with UKTrade Mock SSO testing
mock-sso requires you to provide a `code` param in the initial
GET /o/authorize query string. It passes this param straight back out
during the POST /o/token call. (`code` param is basically used as
a 'fixture' setup for the token call). An app which uses this lib
(django-staff-sso-client) will likely want to be able to test
against mock-sso and so we include this slightly hacky bit of setup
here.