feat: add AUTHBROKER_INTERNAL_URL
This adds the setting AUTHBROKER_INTERNAL_URL. This is the URL that the client
uses internally to fetch and check tokens, i.e. not exposed to the user.

This is primarily for easier local development with
https://github.com/uktrade/mock-sso using Docker compose. While working with
mock-sso is possible without this change, having the internal/external URLs the
same means you have to have a /etc/hosts host pointing to localhost that
matches the mock-sso service name in the compose file. Having them separate
means you don't have to make any /etc/hosts changes.

This should also ease future changes to the client where communication with
Staff SSO on the server side is _not_ over the public internet.